final cleanup and test fixures
