+++ api/src/core/execution/TestSuiteExecutionStage.cs
@@ -40,7 +40,7 @@ namespace GdUnit4.Executions
 
                     if (testCaseContext.IsFailed || testCaseContext.IsError)
                     {
-                        break;
+                        //break;
                     }
 
                 }
