+++ test/project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="GdUnit4MonoTest"
-run/main_scene="res://TestRunner.tscn"
+run/main_scene="res://src/api/TestRunner.tscn"
 config/features=PackedStringArray("4.1", "C#", "Forward Plus")
 config/icon="res://icon.svg"
